allow "premature close" for non-core streams
Fix #11

The timing of "close" events is implementation-specific.  For core
streams, it's reliably possible to determine if a close event is coming
in advance of the stream actually being "done".

However, if a stream doesn't have a _readableState or _writableState
property, then it's impossible to determine if the close is "premature"
or not.